requirements.txt
langgraph
langchain-core
langchain-google-genai
langchain-groq
langchain-community
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
pydantic-settings
pytz
beautifulsoup4
lxml
